Adding code to check if file .place_name exist

+++ sra_client.rb
@@ -1,9 +1,12 @@
 # file client_teste.rb
 require 'socket'
 
+IP = "localhost"
+PORT = 3001
+WITHOUT_PLACE = "none"
+
+
 loop {
-	IP = "localhost"
-	PORT = 3001
 
   server = TCPSocket.open( IP, PORT ) # Connect to the server
 
@@ -12,8 +15,6 @@ loop {
 
   place = msg_server.first.chomp
 
-	WITHOUT_PLACE = "none"
-
   if place != WITHOUT_PLACE
     File.open('.place_name', 'w') do |file|
       file.puts place
@@ -21,11 +22,18 @@ loop {
 
     server.puts ''
   else
-    place_name = File.open( '.place_name', 'r' )
 
-    place_name.each_line do |line|
-      server.puts line.chomp
+    if File.exist?('.place_name')
+      place_name = File.open( '.place_name', 'r' )
+
+      place_name.each_line do |line|
+        server.puts line.chomp
+      end
+    else
+      File.new(".place_name", "w")
+      server.puts ''
     end
+
   end
 
   sleep(0.1)
